Moved header files to include/ because of broken header dependencies
