+++ include/protocolexception.h
