+++ src/passthroughbuffer.h
@@ -18,7 +18,7 @@
 #include <memory>
 #include <cstring>
 #include "endian.h"
-#include "frame.h"
+#include "../include/frame.h"
 
 /**
  *  Set up namespace
